add modified libraries
The Adafruit_SSD1306 library was slimmed down and modified to scale the
64x32 CHIP8 screen to the OLED display with larger pixels. This reduces
the size of the screen buffer from 1 KiB to 256 B. Additionally, a custom
font was created and included in the library. In order to show text, the
screen buffer is misused to store the 225 characters which are translated
to pixel patterns on the fly.
The library was also extended by functionality to only update part of the
screen which speeds up the CHIP8 DRW opcode.

The CHIP8 library contains the CHIP8 virtual machine that loads programs
from the external EEPROM and runs them using the external SPI RAM.
The other libraries were copied without modifications.